Support fractional/decimal material quantities
Real stock counts aren't always whole numbers (half a roll, 2.5
pallets, a metered length) — quantity moves from integer to numeric
on both items and reservations, with the reserve_item RPC and
item_availability view updated to match, so these can be entered and
reserved exactly rather than forced to round to a whole unit.

+++ packages/shared/src/schemas/item.ts
@@ -3,7 +3,7 @@ import { z } from "zod";
 export const itemFormSchema = z.object({
   name: z.string().trim().min(1, "Material name is required"),
   identification_number: z.string().trim().optional().or(z.literal("")),
-  quantity: z.coerce.number().int().min(0, "Quantity cannot be negative"),
+  quantity: z.coerce.number().min(0, "Quantity cannot be negative"),
   // What's being counted (e.g. "bundles", "rolls", "bags") — shown next to
   // the number instead of a bare integer. Optional; blank just shows the
   // number on its own.
